Some improvements.
We don't want to use std::random_device because we need pseudo random
numbers that are the same every time.
std::mt19937 must be a class member, because it has internal state;
it only generates new random numbers when it isn't reconstructed
every time (which is quite expensive, too).
Moved idle() outside of the switch to the end, because it needs
to be called always when the state didn't change (or yield()).